+++ src/multiple-sources.js
@@ -17,15 +17,16 @@ inherit(_AbstractPool, Stream, {
 
   _name: 'abstractPool',
 
-  _add: function(obs) {
+  _add: function(obj, toObs) {
+    toObs = toObs || id;
     if (this._concurLim === -1 || this._curSources.length < this._concurLim) {
-      this._addToCur(obs);
+      this._addToCur(toObs(obj));
     } else {
       if (this._queueLim === -1 || this._queue.length < this._queueLim) {
-        this._addToQueue(obs);
+        this._addToQueue(toObs(obj));
       } else if (this._drop === 'old') {
         this._removeOldest();
-        this._add(obs);
+        this._add(toObs(obj));
       }
     }
   },
@@ -260,7 +261,7 @@ inherit(FlatMap, _AbstractPool, {
   _handleMainSource: function(event) {
     if (event.type === VALUE) {
       if (!event.current || this._lastCurrent !== event.value) {
-        this._add(this._fn(event.value));
+        this._add(event.value, this._fn);
       }
       this._lastCurrent = event.value;
     } else {